feat: break out shareable logic for web
