+++ apps/agentic-chat/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@ai-sdk/react": "^2.0.76",
+    "@shapeshiftoss/chat": "workspace:^",
     "@dynamic-labs/ethereum": "^4.49.0",
     "@dynamic-labs/ethereum-core": "^4.49.0",
     "@dynamic-labs/sdk-react-core": "^4.49.0",
