+++ apps/agentic-chat/src/app/dashboard/page.tsx
@@ -1,14 +1,47 @@
+import { ChatProvider, generateConversationId, useChatContext } from '@shapeshiftoss/chat'
+import { useEffect, useRef } from 'react'
+import { useParams, useNavigate } from 'react-router-dom'
+
 import { Chat } from '@/components/Chat'
 import { ConnectWallet } from '@/components/ConnectWallet'
 import { SidebarLeft } from '@/components/SidebarLeft'
 import { SidebarInset, SidebarProvider, SidebarTrigger } from '@/components/ui/Sidebar'
-import { ChatProvider } from '@/providers/ChatProvider'
+import { useWalletConnection } from '@/hooks/useWalletConnection'
+import { analytics } from '@/lib/mixpanel'
 
 const isSidebarLeftEnabled = import.meta.env.VITE_FEATURE_ENABLE_SIDEBAR_LEFT === 'true'
 
 export const Dashboard = () => {
+  const { conversationId } = useParams<{ conversationId?: string }>()
+  const navigate = useNavigate()
+  const wallet = useWalletConnection()
+
+  // App handles auto-redirect for missing conversation ID
+  useEffect(() => {
+    if (!conversationId) {
+      const newId = generateConversationId()
+      void navigate(`/chats/${newId}`, { replace: true })
+    }
+  }, [conversationId, navigate])
+
   return (
-    <ChatProvider>
+    <ChatProvider
+      conversationId={conversationId}
+      walletState={{
+        evmAddress: wallet.evmAddress,
+        solanaAddress: wallet.solanaAddress,
+        approvedChainIds: wallet.approvedChainIds,
+      }}
+      onConversationDelete={id => {
+        // If deleting active conversation, create new one
+        if (id === conversationId) {
+          const newId = generateConversationId()
+          void navigate(`/chats/${newId}`)
+        }
+      }}
+      apiBaseUrl={import.meta.env.VITE_AGENTIC_SERVER_BASE_URL}
+    >
+      <ChatAnalyticsWrapper />
       <SidebarProvider>
         {isSidebarLeftEnabled && <SidebarLeft />}
         <SidebarInset className="h-dvh flex flex-col">
@@ -25,4 +58,23 @@ export const Dashboard = () => {
       </SidebarProvider>
     </ChatProvider>
   )
+}
+
+// Wrapper that adds analytics to chat events
+function ChatAnalyticsWrapper() {
+  const { messages, status } = useChatContext()
+  const lastTrackedMessageIdRef = useRef<string | null>(null)
+
+  useEffect(() => {
+    // Only track when actively submitting a new message, not when loading history
+    if (status !== 'submitted') return
+
+    const lastMessage = messages[messages.length - 1]
+    if (lastMessage?.role === 'user' && lastMessage.id !== lastTrackedMessageIdRef.current) {
+      analytics.trackChatMessage()
+      lastTrackedMessageIdRef.current = lastMessage.id
+    }
+  }, [messages, status])
+
+  return null
 }
