+++ apps/agentic-chat/vite.config.mjs
@@ -8,7 +8,7 @@ export default defineConfig(() => ({
     cacheDir: resolve(__dirname, '../../node_modules/.vite/agentic-chat'),
     define: {
         global: 'globalThis',
-        'process.env': {},
+        'process.env': '{}',
     },
     server: {
         port: 4200,
