+++ packages/agentic-chat/.npmignore
@@ -1,0 +1,26 @@
+# Source files (only publish dist/)
+src/
+*.ts
+*.tsx
+!*.d.ts
+
+# Config files
+tsconfig.json
+tsconfig.*.json
+
+# Development
+node_modules/
+.DS_Store
+*.log
+**/*.tsbuildinfo
+
+# Tests (if added later)
+__tests__/
+*.test.ts
+*.test.tsx
+*.spec.ts
+*.spec.tsx
+
+# CI/CD
+.github/
+.gitignore
