+++ packages/agentic-chat/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "@shapeshiftoss/chat",
+  "version": "0.1.0",
+  "description": "Internal ShapeShift package for sharing chat infrastructure between shapeshift-agentic and shapeshift web repos. Not designed for external use.",
+  "type": "module",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist",
+    "README.md",
+    "LICENSE"
+  ],
+  "scripts": {
+    "clean": "rm -rf dist",
+    "build": "rm -rf dist && tsc --build",
+    "dev": "tsc --build --watch --preserveWatchOutput",
+    "type-check": "tsc --noEmit",
+    "prepublishOnly": "npm run build"
+  },
+  "keywords": [
+    "chat",
+    "ai",
+    "conversation",
+    "shapeshift",
+    "web3",
+    "crypto",
+    "zustand",
+    "react"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/shapeshift/shapeshift-agentic.git",
+    "directory": "packages/agentic-chat"
+  },
+  "homepage": "https://github.com/shapeshift/shapeshift-agentic#readme",
+  "license": "MIT",
+  "author": "ShapeShift",
+  "dependencies": {
+    "@ai-sdk/react": "^2.0.76",
+    "@ai-sdk/ui-utils": "^1.2.11",
+    "ai": "^5.0.76",
+    "zustand": "^5.0.6",
+    "immer": "^10.1.3"
+  },
+  "peerDependencies": {
+    "react": ">=18.0.0",
+    "react-dom": ">=18.0.0"
+  },
+  "devDependencies": {
+    "@types/react": "*",
+    "@types/react-dom": "*"
+  }
+}
