+++ packages/agentic-chat/src/index.ts
@@ -1,0 +1,19 @@
+// Core provider
+export { ChatProvider, useChatContext } from './providers/ChatProvider'
+
+// Store
+export { useChatStore, createChatStore } from './store/chatStore'
+export { saveMessages, loadMessages, deleteMessages, createMessageStorage } from './store/messageStorage'
+
+// Hooks
+export { useToolExecutionEffect } from './hooks/useToolExecutionEffect'
+
+// Utilities
+export { generateConversationId, extractTitleFromMessages } from './utils/conversationStorage'
+export { createStepPhaseMap, getStepStatus, StepStatus } from './utils/stepUtils'
+
+// Types
+export type { Conversation } from './types/conversation'
+export type { PersistedToolState } from './types/toolState'
+export type { StepState } from './types/stepState'
+export type { Message } from './types/message'
