+++ packages/agentic-chat/src/types/conversation.ts
