+++ packages/agentic-chat/src/types/index.ts
@@ -1,0 +1,4 @@
+export type { Conversation } from './conversation'
+export type { PersistedToolState } from './toolState'
+export type { StepState } from './stepState'
+export type { Message } from './message'
