+++ packages/agentic-chat/src/types/message.ts
@@ -1,0 +1,3 @@
+import type { useChat } from '@ai-sdk/react'
+
+export type Message = ReturnType<typeof useChat>['messages'][number]
