+++ packages/agentic-chat/src/types/stepState.ts
@@ -1,0 +1,6 @@
+export interface StepState<TStep extends number> {
+  currentStep: TStep
+  completedSteps: Set<TStep>
+  failedStep?: TStep
+  error?: string
+}
